pyproject.toml
[project]
name = "multi-tenant-team-task-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.0",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.0",
    "pendulum>=3.1.0",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "ruff>=0.14.11",
]
